ecommerce-admin/backend/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
Flask-Cors==3.0.10
PyJWT==2.7.0
pandas==2.0.2
openpyxl==3.1.2
mysqlclient==2.1.1
python-dotenv==1.0.0
gunicorn==20.1.0
nbformat>=4.2.0
